+++ app/(public)/_components/UserDropdown.tsx
@@ -18,9 +18,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import Link from "next/link";
-import { authClient } from "@/lib/auth-client";
-import { useRouter } from "next/navigation";
-import { toast } from "sonner";
+import { useSignOut } from "@/hooks/use-signout";
 
 interface iAppProps {
   name: string;
@@ -29,21 +27,8 @@ interface iAppProps {
 }
 
 export function UserDropdown({name, email, image} : iAppProps) {
-  const router = useRouter(); // ✅ ĐÚNG - Hook được gọi ở top level của component
-  
-  async function signOut() {
-    await authClient.signOut({
-      fetchOptions: {
-        onSuccess: () => {
-          router.push("/");
-          toast.success("Đăng xuất thành công");
-        },
-        onError: () => {
-          toast.error("Đăng xuất thất bại");
-        },
-      },
-    });
-  }
+ 
+  const handleSignOut = useSignOut();
 
   return (
     <DropdownMenu>
@@ -95,7 +80,7 @@ export function UserDropdown({name, email, image} : iAppProps) {
           </DropdownMenuItem>
         </DropdownMenuGroup>
         <DropdownMenuSeparator />
-        <DropdownMenuItem onClick={signOut}>
+        <DropdownMenuItem onClick={handleSignOut}>
           <LogOutIcon size={16} className="opacity-60" aria-hidden="true" />
           <span>Đăng xuất</span>
         </DropdownMenuItem>
