+++ app/admin/courses/page.tsx
@@ -1,0 +1,7 @@
+export default function CoursesPage() {
+    return (
+        <div>
+            Page courses
+        </div>
+    )
+}
